Updated commit.

+++ Enciclopedie/Controllers/ArticleController.cs
@@ -63,22 +63,20 @@ namespace Enciclopedie.Controllers
         [NonAction]
         public IEnumerable<SelectListItem> GetAllCategories()
         {
-            // generam o lista goala
             var selectList = new List<SelectListItem>();
-            // Extragem toate categoriile din baza de date
+           
             var categories = from cat in db.Categories
                              select cat;
-            // iteram prin categorii
+           
             foreach (var category in categories)
             {
-                // Adaugam in lista elementele necesare pentru dropdown
                 selectList.Add(new SelectListItem
                 {
                     Value = category.CategoryId.ToString(),
                     Text = category.CategoryName.ToString()
                 });
             }
-            // returnam lista de categorii
+            
             return selectList;
         }
 
@@ -173,7 +171,11 @@ namespace Enciclopedie.Controllers
             Image image = db.Images.Find(id);
             if (article.UserId == User.Identity.GetUserId() || User.IsInRole("Administrator"))
             {
-                db.Images.Remove(image);
+                if(image!=null)
+                {
+                    db.Images.Remove(image);
+                }
+              
                 db.Articles.Remove(article);
                 db.SaveChanges();
                 TempData["message"] = "Articolul a fost sters!";
